🔧 fix: SideSection 페이지 이동
