--- src/components/NavBar.jsx
@@ -1,3 +1,0 @@
-const NavBar = () => {};
-
-export default NavBar;
